# Faster and fixed big.Int -> uint256.Int conversion