@holiman 2020-03-15
heh
```
SetFromBig/1word-6       29.3ns ± 0%   3.4ns ± 1%  -88.37%  (p=0.008 n=5+5)
SetFromBig/2words-6      48.3ns ± 0%   3.7ns ± 0%  -92.35%  (p=0.000 n=5+4)
SetFromBig/3words-6      66.8ns ± 0%   4.0ns ± 0%  -94.04%  (p=0.008 n=5+5)
SetFromBig/4words-6      80.6ns ± 0%   4.2ns ± 0%  -94.73%  (p=0.000 n=5+4)
SetFromBig/overflow-6    96.5ns ± 0%   4.2ns ± 0%  -95.60%  (p=0.008 n=5+5)
```
That's pretty sweet :)
:shipit: