@chfast 2020-03-15
> heh
> 
> ```
> SetFromBig/1word-6       29.3ns ± 0%   3.4ns ± 1%  -88.37%  (p=0.008 n=5+5)
> SetFromBig/2words-6      48.3ns ± 0%   3.7ns ± 0%  -92.35%  (p=0.000 n=5+4)
> SetFromBig/3words-6      66.8ns ± 0%   4.0ns ± 0%  -94.04%  (p=0.008 n=5+5)
> SetFromBig/4words-6      80.6ns ± 0%   4.2ns ± 0%  -94.73%  (p=0.000 n=5+4)
> SetFromBig/overflow-6    96.5ns ± 0%   4.2ns ± 0%  -95.60%  (p=0.008 n=5+5)
> ```
> 
> That's pretty sweet :)
> :shipit:

In the end your implementation turned out to be the fastest (at least for these benchmarks where branch prediction should be perfect because number of works is always the same).
Interestingly, it is better to use `if len(words) > ...` than `if numWords > ...`. In the second one bounds checking is added for `words` access.

The switch is resolved at compile time - only code for one case is emitted. For 32-bit case I didn't bother to unroll it, but this can be done in the same way as for 64-bit case.